+++ packages/remix-server-runtime/cookies.ts
@@ -1,12 +1,16 @@
-import type { CookieParseOptions, CookieSerializeOptions } from "cookie";
+import type {
+  CookieParseOptions as RRCookieParseOptions,
+  CookieSerializeOptions as RRCookieSerializeOptions,
+} from "cookie";
 import { parse, serialize } from "cookie";
 
 import type { SignFunction, UnsignFunction } from "./crypto";
 import { warnOnce } from "./warnings";
 
-export type { CookieParseOptions, CookieSerializeOptions };
+export type CookieParseOptions = RRCookieParseOptions;
+export type CookieSerializeOptions = RRCookieSerializeOptions;
 
-export interface CookieSignatureOptions {
+export type CookieSignatureOptions = {
   /**
    * An array of secrets that may be used to sign/unsign the value of a cookie.
    *
@@ -16,7 +20,7 @@ export interface CookieSignatureOptions {
    * cookies that were signed with older secrets still work.
    */
   secrets?: string[];
-}
+};
 
 export type CookieOptions = CookieParseOptions &
   CookieSerializeOptions &
@@ -32,7 +36,7 @@ export type CookieOptions = CookieParseOptions &
  *
  * @see https://remix.run/utils/cookies#cookie-api
  */
-export interface Cookie {
+export type Cookie = {
   /**
    * The name of the cookie, used in the `Cookie` and `Set-Cookie` headers.
    */
@@ -65,7 +69,7 @@ export interface Cookie {
    * header.
    */
   serialize(value: any, options?: CookieSerializeOptions): Promise<string>;
-}
+};
 
 export type CreateCookieFunction = (
   name: string,