+++ packages/remix-server-runtime/routes.ts
@@ -8,35 +8,35 @@ import { callRouteActionRR, callRouteLoaderRR } from "./data";
 import type { FutureConfig } from "./entry";
 import type { ServerRouteModule } from "./routeModules";
 
-export interface RouteManifest<Route> {
+export type RouteManifest<Route> = {
   [routeId: string]: Route;
-}
+};
 
 export type ServerRouteManifest = RouteManifest<Omit<ServerRoute, "children">>;
 
 // NOTE: make sure to change the Route in remix-react if you change this
-export interface Route {
+export type Route = {
   index?: boolean;
   caseSensitive?: boolean;
   id: string;
   parentId?: string;
   path?: string;
-}
+};
 
 // NOTE: make sure to change the EntryRoute in remix-react if you change this
-export interface EntryRoute extends Route {
+export type EntryRoute = Route & {
   hasAction: boolean;
   hasLoader: boolean;
   hasErrorBoundary: boolean;
   imports?: string[];
   module: string;
   parentId?: string;
-}
+};
 
-export interface ServerRoute extends Route {
+export type ServerRoute = Route & {
   children: ServerRoute[];
   module: ServerRouteModule;
-}
+};
 
 function groupRoutesByParentId(manifest: ServerRouteManifest) {
   let routes: Record<string, Omit<ServerRoute, "children">[]> = {};